feat(): create script structure
Created the structure of the python script and divided it into separate tasks.
- get_html returns HTML page string from its url. It uses requests,
which is an additional package I need to ask approval for.

- get_paragraphs returns text paragraphs from an HTML page string

- classify returns the classification of a paragraph - ingredient,
instruction or irrelevant, coded 1, 0, -1. This seems to be
the main task of this assignment, and I will learn a bit of NLP for it

- get_recipe_json uses the functions above to compose the apropriate
output json from an input URL.