feat(model): add model and training
FUNCTION:
From a new model.py file, can now use train_model function to train a neural network to classify between ingredients and instructions, and use predict to utilize the model on new information. However, the saved model does not currently work properly (see notes)

METHOD:
- Parse the tsv data into data and label lists

- Separated into training data (~80k entries) and testing data (~30k entries)

- Created a tokenizer with a word index based on the training data, used it to sequence the training data and then padded the result

- Created a model with an Embedding layer and an average pooling layer and trained it based on the training data until it rached sufficient accuracy (99%) with the testing data, and then saved it
- The trained model also proves accurate on entries from the example page in the assignment description

NOTES:
While the model seems to save and load succefully, when it is loaded not from the same session it was trained in it seems to behave completely differently.

 It still recognizes instructions with high confidence but gives everything a high grade, maybe except things it didn't encounter at all prior. 

This could be a case of https://github.com/keras-team/keras/issues/4875 , to which I have not found a working solution. 

It may also be due to the fact I am trying to use a single output neuron, while the result is not really binary (ingredient != not instruction). Evidence for this are the fact that it still recognizes instructions ('1') very well and that it is overtly confident in input that fits nither.