feat(data): add a script to help me sort through webpage data
FUNCTION:
running sort-data.sh with a recipe URL as an argument. will open a temporary file with the webpage text content. I can then cut imgredients and instructions (possibly keeping them as extra data in ingredients.txt and instructions.txt here), and save and close the file. Then the script will automatically append the remaining data into neither.txt in the project's directory

METHOD:
Script is written in bash.
- Evaluate python functions from main to get the list of paragraphs from the given web page
- Write them (newline joined) into /tmp/neither
- Open emacs editor on /tmp/neither
- I can cut and paste relevant lines into ingredients.txt and instructions.txt via PyCharm
- The script outputs a message before opening emacs, and waits until emacs is closed.
- After so, it copies the final contents of /tmp/neither into the end of neither.txt

NOTES:
The main reason for this script is the fact that I need training data for all of the irrelevant junk that plagues the data I recieve from the web pages. This mimics the process my scripy goes through, letting me do the job of the classifier, making this quite authentric and also much faster than trying to just do it by hand.