feat(data): get_data gathers from text files as well
FUNCTION:
get_data.py assembles the text files from sort-data.sh into the TSV file together with the JSON. neither data is asssembled partially in the start and partially in the end, for both training and testing data

METHOD:
- in get_data_lists() which gathered the data from the JSON, read the relevant files and add them to the relevant list. Also add a list for irrelevants in the return statement.

- in save_to_tsv(), which writes the data to the TSV file, I also iterated over neither data. 20000 of it (currently 25000) went in the start and the rest in the end. This is because the final data is cut in the middle for training and testing data.

NOTES:
I can't iterate over the neither dtaa together with instructions and ingredients, since it's much smaller. Because of this, I can't alternate with them elegantly when writing to the file so I went with the append/prepend solution, though it's somewhat inelegant as well.