chore(model): divide to classifier and trash detector. Improve cleanup
FUNCTION:
The structure is now thus:
A "Classifier" model calculates the main classification, instruction or ingredient
A "Trash Detector" model (nicknamed "TrashBot") calculates the probabilty of a text being website junk

The trash detector is trained on both the recipe dataset and the junk dataset to produce binary output. Having no other patterns to check by it's incentivized to be less insecure than the previous 3-way classification. 

Once the Trash Detector has confirmed that a text is not junk, the classifier can do an accurate job of detecting what kind of not-junk it is.

All of this improves accuracy by a lot.

In addition, due to my commiting schedule, this includes  vastly improved version of the text cleanup. In a new utilities class, I use a filter to exclude both non-text elements and useless lines (e.g ' ' and ', ').Then I use a map to reduce tabs, newlines and multiple spaces into a single space.

METHOD:

- Most of the model methods (loading, saving, preprocessing) are moved to a new file called nlp_utils.py, and are either wrapped or used directly form there in the old model file.
- assemble_data ignores neither file, and puts only ingredients and onstructions in the TSV.
-  A new trashbot.py is created, including a model with a single binary output using binary crossentopy. It is trained on the TSV and the neither data. Otherwise it is identical to the old model, with paths named apropriately.
- The old model is called classifier  to distinguish it from the trash bot, and file paths are changed accordingly.