bug(model): classifier lost accuracy, since didn't remove duplicates
BUG:
Assembling the data and training the model resulted in singificantly worse results, staying far from the previously-working 99% threshold to the point of an unqualified final product.

CAUSE:
It seems, that when cleaning up assemble_data, retrieval of scraped data was not  stripped of duplicates, resulting in a worse dataset overall. 

FIX: 
Clear scraped data lists of duplicate via `list(set(` constructors

NOTES:
- The 0.99 line still does not work, for some arcane reason. Threshold was lowered to 97% as result.